Refactored RiderAdapter to RiderListAdapter
