+++ app/src/main/res/layout/layout_item_rider.xml
@@ -15,6 +15,7 @@
             android:layout_height="wrap_content"
             android:layout_centerHorizontal="true"
             android:padding="2dp"
+            android:layout_marginBottom="1dp"
             android:textSize="12sp" />
 
     </RelativeLayout>
